Adding donation page and platform
