+++ client/src/components/pages/SignUp.js
@@ -1,0 +1,99 @@
+import React, { useState } from 'react';
+import { Link } from 'react-router-dom';
+import { useMutation } from '@apollo/client';
+import Auth from '../../utils/auth';
+import { ADD_USER } from '../../utils/mutations';
+
+function Signup(props) {
+  const [formState, setFormState] = useState({ email: '', password: '' });
+  const [addUser] = useMutation(ADD_USER);
+
+  const handleFormSubmit = async (event) => {
+    event.preventDefault();
+    const { data } = await addUser({
+      variables: {
+        email: formState.email,
+        password: formState.password,
+      },
+    });
+    console.log(data)
+    const token = data.createUser.token;
+    Auth.login(token);
+  };
+
+  const handleChange = (event) => {
+    const { name, value } = event.target;
+    setFormState({
+      ...formState,
+      [name]: value,
+    });
+  };
+
+// const Signup = () => {
+//   const [formState, setFormState] = useState({
+//     email: '',
+//     password: '',
+//   });
+//   const [addUser] = useMutation(ADD_USER);
+
+//   // update state based on form input changes
+//   const handleChange = (event) => {
+//     const { name, value } = event.target;
+
+//     setFormState({
+//       ...formState,
+//       [name]: value,
+//     });
+//   };
+
+//   const handleFormSubmit = async (event) => {
+//     event.preventDefault();
+//     console.log(formState);
+
+//     try {
+//       const { data } = await addUser({
+//         variables: { ...formState },
+//       });
+
+//       Auth.login(data.addUser.token);
+//     } catch (e) {
+//       console.error(e);
+//     }
+//   };
+
+
+  return (
+    <div className="container my-1">
+      <Link to="/login">← Go to Login</Link>
+
+      <h2>Signup</h2>
+      <form onSubmit={handleFormSubmit}>
+        <div className="flex-row space-between my-2">
+          <label htmlFor="email">Email:</label>
+          <input
+            placeholder="youremail@test.com"
+            name="email"
+            type="email"
+            id="email"
+            onChange={handleChange}
+          />
+        </div>
+        <div className="flex-row space-between my-2">
+          <label htmlFor="pwd">Password:</label>
+          <input
+            placeholder="******"
+            name="password"
+            type="password"
+            id="pwd"
+            onChange={handleChange}
+          />
+        </div>
+        <div className="flex-row flex-end">
+          <button type="submit">Submit</button>
+        </div>
+      </form>
+    </div>
+  );
+}
+
+export default Signup;
