+++ server/utils/auth.js
@@ -1,10 +1,10 @@
 const jwt = require("jsonwebtoken");
 const secret = "secretsecretboi";
-const expiration = "2h";
+const expiration = "5h";
 
 module.exports = {
   authMiddleware: function ( {req} ) {
-    let token = req.query.token || req.headers.authorization;
+    let token = req.body.token || req.query.token || req.headers.authorization;
     if (req.headers.authorization) {
       token = token.split(" ").pop().trim();
     }
